requirements.txt
selenium
pytest
webdriver_manager
faker
allure-pytest
